refactor: 콘텐츠 조회, 실시간 시청 세션 및 스포츠 배치 구조 개선

+++ src/main/java/com/codeit/mople/domain/playlist/dto/response/PlaylistContentResponse.java
@@ -13,7 +13,7 @@ public record PlaylistContentResponse(
     String thumbnailUrl,
     List<String> tags,
     double averageRating,
-    int reviewCount
+    long reviewCount
 ) {
 
   public static PlaylistContentResponse from(PlaylistContent playlistContent) {
